% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_ce.R
\name{effect_ce}
\alias{effect_ce}
\title{STATISTICAL EFFECTS}
\usage{
effect_ce(p0_e1, p0_e2, p1_e1, p1_e2, rho, effect_ce = "diff")
}
\arguments{
\item{p0_e1}{numeric parameter, probability of occurrence E1 in the control group}

\item{p0_e2}{numeric parameter, probability of occurrence E2 in the control group}

\item{p1_e1}{numeric parameter, probability of occurrence E1 in the intervention group}

\item{p1_e2}{numeric parameter, probability of occurrence E2 in the intervention group}

\item{rho}{numeric parameter, Pearson correlation between E1 and E2}

\item{effect_ce}{character, specifies the type of effect measure to be calculated
(effect_ce = "diff" for difference of proportions, effect_ce = "rr" for risk ratio, effect_ce = "or" for odds ratio)}
}
\value{
Returns the desired effect of the composite binary event and the effects of the events
}
\description{
This function calculates different effect measures for binary composite outcomes. In particular,
it allows for computing the risk difference, risk ratio and odds ratio.
}
\details{
The input parameters represent the probability of the composite components and Pearson's correlation between the two components.
Note that Pearson's correlation takes values between two bounds that depend on the probabilities p0_e1 and p0_e2.
To calculate the correlation bounds you can use the R functions lower_corr and upper_corr, available in this package.
}
