% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_corr.R
\name{lower_corr}
\alias{lower_corr}
\title{LOWER LIMIT FOR PEARSON CORRELATION}
\usage{
lower_corr(p_e1, p_e2)
}
\arguments{
\item{p_e1}{numeric parameter, probability of the event E1}

\item{p_e2}{numeric parameter, probability of the event E2}
}
\value{
Returns the lower correlation threshold that the binominal compost event can reach
}
\description{
Knowing that the correlation of a combined variable is presented in an intervalic way,
and has been calculated from the probability of occurrence of the respective events.
This function allows to calculate the minimum correlation
that this typology of variables can experience.
}
\details{
lower_corr returns a numeric value negated bounded between -1 and 0.
The probabilities of the occurrence of events must be defined by the open interval of (0.1).
}
\references{
M.Bofill , G.Gomez , A new approach for sizing trials with composite binary endpoints using anticipated marginal values and accounting for the correlation between components, http://cinna.upc.edu:3838/compare/compareCover/
}
\author{
Raquel Rovira Salvat
}
