% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_ce.R
\name{prob_ce}
\alias{prob_ce}
\title{PROBABILITY OF UNION OF TWO EVENTS}
\usage{
prob_ce(p_e1, p_e2, rho)
}
\arguments{
\item{p_e1}{numeric parameter, probability of the event E1}

\item{p_e2}{numeric parameter, probability of the event E2}

\item{rho}{numeric parameter, Pearson correlation between E1 i E2}
}
\value{
Returns the probability of a union of two events
}
\description{
This function is used to calculate the probability of the union of two events, E1 and E2,
knowing in advance the probability of occurrence of each of the two events by separating
them on the basis of the study population, and in the other band knowing
the Pearson coefficient that correlates the events.
}
\details{
Returns a numeric value. The input parameters representing the probability of the events
taking place are limited between 0 and 1, without including both values. Pearson's correlation
must be within the confidence interval that allows the combined variable according to the probabilities given.
To calculate this confidence interval you can use lower_corr and upper_corr functions that you can find in this package.
}
