% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_corr.R
\name{upper_corr}
\alias{upper_corr}
\title{UPPER LIMIT FOR PEARSON CORRELATION}
\usage{
upper_corr(p_e1, p_e2)
}
\arguments{
\item{p_e1}{numeric parameter, probability of the event E1}

\item{p_e2}{numeric parameter, probability of the event E2}
}
\value{
Returns the upper correlation threshold that can be set by the compost binarius event.
}
\description{
Knowing that the correlation of a combined variable is presented in an intervalic manner,
and calculated from the probability of occurrence of the respective events.
This function allows to calculate the minimum correlation that this typology of variables can experience.
}
\details{
upper_corr returns a numeric value negated bounded between 0 and 1.
The probabilities of the occurrence of events must be defined by the open interval of (0.1).
}
